Add more functionality for field classes

+++ backstaff/fields.py
@@ -1,4 +1,9 @@
+from os import path
+import pathlib
+import glob
+import re
 import numpy as np
+import scipy.interpolate
 try:
     import backstaff.plotting as plotting
 except ModuleNotFoundError:
@@ -25,6 +30,9 @@ class Coords2:
         all_coords.pop(omitted_axis)
         return Coords2(all_coords[0], all_coords[1])
 
+    def get_shape(self):
+        return self.x.size, self.y.size, self.z.size
+
     def __init__(self, x_coords, y_coords):
         self.x = np.asfarray(x_coords)
         self.y = np.asfarray(y_coords)
@@ -41,11 +49,86 @@ class Coords1:
         return self.coords.size
 
 
+class ScalarField3:
+    @staticmethod
+    def from_bifrost_data(
+        bifrost_data,
+        quantities,
+        height_range=None,
+        scale=None,
+        value_processor=lambda x: x,
+    ):
+        z_coords = -(2*bifrost_data.z - bifrost_data.zdn)[::-1]
+        if not isinstance(quantities, list) and not isinstance(
+                quantities, tuple):
+            quantities = [quantities]
+        k_slice = slice(*((None, ) if height_range is None else np.
+                          searchsorted(z_coords, height_range)))
+        values = value_processor(
+            *(bifrost_data.get_var(quantity)[:, :, ::-1][400, 400, k_slice]
+              for quantity in quantities))
+        z_coords = z_coords[k_slice]
+        if scale is not None:
+            values *= scale
+
+        return ScalarField1(Coords1(z_coords), values)
+        return ScalarField3(Coords3(bifrost_data.x, bifrost_data.y, z_coords),
+                            values)
+
+    def __init__(self, coords, values):
+        assert isinstance(coords, Coords3)
+        self.coords = coords
+        self.values = np.asfarray(values)
+        assert self.values.shape == self.coords.get_shape()
+
+    def __add__(self, term):
+        if isinstance(term, self.__class__):
+            assert np.allclose(self.coords.x, term.coords.x)
+            assert np.allclose(self.coords.y, term.coords.y)
+            return ScalarField3(self.coords, self.values + term.values)
+        else:
+            return ScalarField3(self.coords, self.values + term)
+
+    def __sub__(self, term):
+        if isinstance(term, self.__class__):
+            assert np.allclose(self.coords.x, term.coords.x)
+            assert np.allclose(self.coords.y, term.coords.y)
+            return ScalarField3(self.coords, self.values - term.values)
+        else:
+            return ScalarField3(self.coords, self.values - term)
+
+    def __mul__(self, factor):
+        if isinstance(factor, self.__class__):
+            assert np.allclose(self.coords.x, factor.coords.x)
+            assert np.allclose(self.coords.y, factor.coords.y)
+            return ScalarField3(self.coords, self.values*factor.values)
+        else:
+            return ScalarField3(self.coords, self.values*factor)
+
+    def __truediv__(self, divisor):
+        if isinstance(divisor, self.__class__):
+            assert np.allclose(self.coords.x, divisor.coords.x)
+            assert np.allclose(self.coords.y, divisor.coords.y)
+            return ScalarField3(self.coords, self.values/divisor.values)
+        else:
+            return ScalarField3(self.coords, self.values/divisor)
+
+    def get_shape(self):
+        return self.coords.get_shape()
+
+    def get_values(self):
+        return self.values
+
+    def get_values_flat(self):
+        return np.ravel(self.values)
+
+
 class ScalarField1:
     @staticmethod
     def horizontal_average_from_bifrost_data(
         bifrost_data,
         quantities,
+        height_range=None,
         scale=None,
         value_processor=lambda x: x,
     ):
@@ -53,10 +136,13 @@ class ScalarField1:
         if not isinstance(quantities, list) and not isinstance(
                 quantities, tuple):
             quantities = [quantities]
+        k_slice = slice(*((None, ) if height_range is None else np.
+                          searchsorted(coords, height_range)))
         all_values = value_processor(
-            *(bifrost_data.get_var(quantity)[:, :, ::-1]
+            *(bifrost_data.get_var(quantity)[:, :, ::-1][:, :, k_slice]
               for quantity in quantities))
         values = np.mean(all_values, axis=(0, 1))
+        coords = coords[k_slice]
         if scale is not None:
             values *= scale
 
@@ -68,6 +154,7 @@ class ScalarField1:
         i,
         j,
         quantities,
+        height_range=None,
         scale=None,
         value_processor=lambda x: x,
     ):
@@ -75,10 +162,40 @@ class ScalarField1:
         if not isinstance(quantities, list) and not isinstance(
                 quantities, tuple):
             quantities = [quantities]
-        values = value_processor(*(bifrost_data.get_var(quantity)[i, j, ::-1]
-                                   for quantity in quantities))
+        k_slice = slice(*((None, ) if height_range is None else np.
+                          searchsorted(coords, height_range)))
+        values = value_processor(
+            *(bifrost_data.get_var(quantity)[i, j, ::-1][k_slice]
+              for quantity in quantities))
+        coords = coords[k_slice]
         if scale is not None:
             values *= scale
+
+        return ScalarField1(Coords1(coords), values)
+
+    @staticmethod
+    def dz_in_bifrost_data(bifrost_data, height_range=None, scale=None):
+        coords = -(2*bifrost_data.z - bifrost_data.zdn)[::-1]
+        k_slice = slice(*((None, ) if height_range is None else np.
+                          searchsorted(coords, height_range)))
+        values = (bifrost_data.z - bifrost_data.zdn)[::-1][k_slice]
+        coords = coords[k_slice]
+        scale = 2 if scale is None else (2*scale)
+        values *= scale
+
+        return ScalarField1(Coords1(coords), values)
+
+    @staticmethod
+    def volumes_in_bifrost_data(bifrost_data, height_range=None, scale=None):
+        coords = -(2*bifrost_data.z - bifrost_data.zdn)[::-1]
+        dx = bifrost_data.params['dx'][0]
+        dy = bifrost_data.params['dy'][0]
+        k_slice = slice(*((None, ) if height_range is None else np.
+                          searchsorted(coords, height_range)))
+        values = (bifrost_data.z - bifrost_data.zdn)[::-1][k_slice]
+        coords = coords[k_slice]
+        scale = (2*dx*dy) if scale is None else (2*dx*dy*scale)
+        values *= scale
 
         return ScalarField1(Coords1(coords), values)
 
@@ -121,11 +238,22 @@ class ScalarField1:
         else:
             return ScalarField1(self.coords, self.values/divisor)
 
+    def resampled_to_coords(self, coords, kind='cubic'):
+        values = scipy.interpolate.interp1d(self.get_coords(),
+                                            self.get_values(),
+                                            kind=kind,
+                                            copy=False,
+                                            fill_value='extrapolate')(coords)
+        return ScalarField1(Coords1(coords), values)
+
     def get_size(self):
         return self.coords.get_size()
 
     def get_values(self, inverted_vertically=False):
         return self.values[::(-1 if inverted_vertically else 1)]
+
+    def get_values_flat(self, **kwargs):
+        return self.get_values(**kwargs)
 
     def get_coords(self, inverted=False):
         return -self.coords.coords[::-1] if inverted else self.coords.coords
@@ -153,13 +281,16 @@ class ScalarField1:
             self.get_values(inverted_vertically=inverted_vertically),
             **plot_kwargs)
 
-    def save(self, file_path, compressed=True):
-        if compressed:
-            np.savez_compressed(file_path,
-                                coords=self.coords.coords,
-                                values=self.values)
-        else:
-            np.savez(file_path, coords=self.coords.coords, values=self.values)
+    def save(self, file_path, compressed=True, overwrite=True):
+        if overwrite or not pathlib.Path(file_path).exists():
+            if compressed:
+                np.savez_compressed(file_path,
+                                    coords=self.coords.coords,
+                                    values=self.values)
+            else:
+                np.savez(file_path,
+                         coords=self.coords.coords,
+                         values=self.values)
 
 
 class ScalarField2:
@@ -170,27 +301,47 @@ class ScalarField2:
 
     @staticmethod
     def slice_from_bifrost_data(bifrost_data,
-                                quantity,
+                                quantities,
                                 slice_axis=1,
                                 slice_coord=7.5,
-                                scale=None):
-        all_center_coords = [
-            bifrost_data.x, bifrost_data.y, -bifrost_data.z[::-1]
-        ]
-        slice_idx = np.argmin(
-            np.abs(all_center_coords[slice_axis] - slice_coord))
+                                slice_idx=None,
+                                scale=None,
+                                value_processor=lambda x: x):
+        if slice_idx is None:
+            all_center_coords = [
+                bifrost_data.x, bifrost_data.y, -bifrost_data.z[::-1]
+            ]
+            slice_idx = np.argmin(
+                np.abs(all_center_coords[slice_axis] - slice_coord))
 
         all_slices = [slice(None)]*3
         all_slices[slice_axis] = slice_idx
 
-        values = bifrost_data.get_var(quantity)[:, :, ::-1][all_slices[0],
-                                                            all_slices[1],
-                                                            all_slices[2]]
-        if scale is not None:
-            values = values*scale
+        if not isinstance(quantities, list) and not isinstance(
+                quantities, tuple):
+            quantities = [quantities]
 
-        return ScalarField2(
-            Coords2.from_bifrost_data(bifrost_data, slice_axis), values)
+        values = value_processor(
+            *(bifrost_data.get_var(quantity)[:, :, ::-1][all_slices[0],
+                                                         all_slices[1],
+                                                         all_slices[2]]
+              for quantity in quantities))
+        coords = Coords2.from_bifrost_data(bifrost_data, slice_axis)
+
+        if isinstance(values, dict):
+            if scale is not None:
+                for name in values:
+                    values[name] *= scale
+
+            return ScalarFieldSet(
+                **
+                {name: ScalarField2(coords, v)
+                 for name, v in values.items()})
+        else:
+            if scale is not None:
+                values *= scale
+
+            return ScalarField2(coords, values)
 
     @staticmethod
     def accumulated_from_bifrost_data(bifrost_data,
@@ -202,18 +353,43 @@ class ScalarField2:
         if not isinstance(quantities, list) and not isinstance(
                 quantities, tuple):
             quantities = [quantities]
-        all_values = value_processor(
-            *(bifrost_data.get_var(quantity)[:, :, ::-1]
-              for quantity in quantities))
-        values = accum_operator(all_values, axis=accum_axis)
-        if scale is not None:
-            values *= scale
 
-        return ScalarField2(
-            Coords2.from_bifrost_data(bifrost_data, accum_axis), values)
+        all_values = value_processor(*(bifrost_data.get_var(quantity)
+                                       for quantity in quantities))
+
+        values = accum_operator(all_values, axis=accum_axis)
+        coords = Coords2.from_bifrost_data(bifrost_data, accum_axis)
+
+        if isinstance(values, dict):
+            if accum_axis != 2:
+                for name, v in values.items():
+                    values[name] = v[:, ::-1]
+
+            if scale is not None:
+                for name in values:
+                    values[name] *= scale
+
+            return ScalarFieldSet(
+                **
+                {name: ScalarField2(coords, v)
+                 for name, v in values.items()})
+        else:
+            if accum_axis != 2:
+                values = values[:, ::-1]  # Flip z
+
+            if scale is not None:
+                values *= scale
+
+            return ScalarField2(coords, values)
 
     @staticmethod
-    def from_file(file_path):
+    def from_file(file_path, field_name_in_set=None):
+        try:
+            if field_name_in_set is not None:
+                return ScalarFieldSet.field_from_file(ScalarField2, file_path,
+                                                      field_name_in_set)
+        except FileNotFoundError:
+            pass
         data = np.load(file_path)
         return ScalarField2(Coords2(data['x'], data['y']), data['values'])
 
@@ -292,27 +468,79 @@ class ScalarField2:
                       np.diff(self.get_vertical_coords()))
 
     def plot(self, inverted_vertically=False, **plot_kwargs):
-        figure_width = plot_kwargs.pop('figure_width', 7.2)
-        figure_aspect = plot_kwargs.pop(
-            'figure_aspect', 5/4 if np.abs(
-                (self.get_horizontal_extent() - self.get_vertical_extent())/
-                self.get_horizontal_extent()) < 1e-3 else 4.5/3)
+        aspect_ratio = 5/4 if np.abs(
+            (self.get_horizontal_extent() - self.get_vertical_extent())/
+            self.get_horizontal_extent()) < 1e-3 else 4.5/3
+        fig_kwargs = plot_kwargs.pop('fig_kwargs', {})
+        fig_kwargs['width'] = fig_kwargs.pop('width', 7.2)
+        fig_kwargs['aspect_ratio'] = fig_kwargs.pop('aspect_ratio',
+                                                    aspect_ratio)
         return plotting.plot_2d_field(
             self.get_horizontal_coords(),
             self.get_vertical_coords(inverted=inverted_vertically),
             self.get_values(inverted_vertically=inverted_vertically),
-            figure_width=figure_width,
-            figure_aspect=figure_aspect,
+            fig_kwargs=fig_kwargs,
             **plot_kwargs)
 
-    def save(self, file_path, compressed=True):
-        if compressed:
-            np.savez_compressed(file_path,
-                                x=self.coords.x,
-                                y=self.coords.y,
-                                values=self.values)
-        else:
-            np.savez(file_path,
-                     x=self.coords.x,
-                     y=self.coords.y,
-                     values=self.values)
+    def save(self, file_path, compressed=True, overwrite=True):
+        if overwrite or not pathlib.Path(file_path).exists():
+            if compressed:
+                np.savez_compressed(file_path,
+                                    x=self.coords.x,
+                                    y=self.coords.y,
+                                    values=self.values)
+            else:
+                np.savez(file_path,
+                         x=self.coords.x,
+                         y=self.coords.y,
+                         values=self.values)
+
+
+class ScalarFieldSet:
+    @staticmethod
+    def field_from_file(field_class, base_file_path, field_name):
+        glob_file_path = ScalarFieldSet._get_field_path(base_file_path,
+                                                        '*',
+                                                        escaper=glob.escape)
+        re_file_path = ScalarFieldSet._get_field_path(base_file_path,
+                                                      '(.+)',
+                                                      escaper=re.escape)
+        for file_path in glob.iglob(glob_file_path):
+            match = re.search(re_file_path, file_path)
+            if match is not None:
+                name = match.groups(1)[0]
+                if name == field_name:
+                    field_file_path = ScalarFieldSet._get_field_path(
+                        base_file_path, name)
+                    return field_class.from_file(field_file_path)
+
+        raise FileNotFoundError(
+            f'No field named {field_name} exists for base path {base_file_path}'
+        )
+
+    def __init__(self, **fields):
+        assert len(fields) > 0
+        self.fields = fields
+        self.field_class = next(iter(self.fields.values())).__class__
+        for v in self.fields.values():
+            assert isinstance(v, self.field_class)
+
+    def save(self, file_path, **kwargs):
+        for name, field in self.fields.items():
+            field.save(self.__class__._get_field_path(file_path, name),
+                       **kwargs)
+
+    def __getitem__(self, name):
+        return self.fields[name]
+
+    def __setitem__(self, name, field):
+        assert isinstance(field, self.field_class)
+        self.fields[name] = field
+
+    @staticmethod
+    def _get_field_path(base_file_path,
+                        name,
+                        separator='.',
+                        escaper=lambda x: x):
+        fp = pathlib.Path(base_file_path)
+        return f'{escaper(str(fp.parent))}{escaper("" if fp.parent == "/" else "/")}{escaper(fp.stem)}{escaper(separator)}{name}{escaper(fp.suffix)}'
